pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alert-alchemy"
version = "0.2.0"
description = "A terminal/CLI incident-response simulator game where you 'brew' fixes from noisy alerts"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Alert Alchemy Contributors"}
]
keywords = ["incident-response", "sre", "cli", "game", "simulation", "on-call"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Games/Entertainment :: Simulation",
    "Topic :: System :: Systems Administration",
    "Typing :: Typed",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
gui = [
    "pywebview>=4.0.0",
]

[project.scripts]
alert-alchemy = "alert_alchemy.cli:app"
aa = "alert_alchemy.cli:app"
alert-alchemy-gui = "alert_alchemy.gui_launcher:main"

[project.urls]
Homepage = "https://github.com/ThanhNguyxn/alert-alchemy"
Repository = "https://github.com/ThanhNguyxn/alert-alchemy"
Issues = "https://github.com/ThanhNguyxn/alert-alchemy/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
